% Auto-generated: do not edit by hand
\name{''DateTimePicker}

\alias{''DateTimePicker}

\title{DateTimePicker component}

\description{

}

\usage{
''DateTimePicker(datetime=NULL, renderTimezone=NULL, dateInputStyle=NULL,
timezoneInputStyle=NULL, id=NULL, value=NULL)
}

\arguments{
\item{datetime}{Character. }

\item{renderTimezone}{Logical. }

\item{dateInputStyle}{Named list. }

\item{timezoneInputStyle}{Named list. }

\item{id}{Character. }

\item{value}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

