% Auto-generated: do not edit by hand
\name{''TagInput}

\alias{''TagInput}

\title{TagInput component}

\description{

}

\usage{
''TagInput(wrapperStyle=NULL, tagStyle=NULL, inputStyle=NULL,
tagDeleteStyle=NULL, placeholder=NULL, injectedTags=NULL,
id=NULL, value=NULL)
}

\arguments{
\item{wrapperStyle}{Named list. }

\item{tagStyle}{Named list. }

\item{inputStyle}{Named list. }

\item{tagDeleteStyle}{Named list. }

\item{placeholder}{Character. }

\item{injectedTags}{Unnamed list. }

\item{id}{Character. }

\item{value}{Unnamed list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

