% Auto-generated: do not edit by hand
\name{keyedFileBrowser}

\alias{keyedFileBrowser}

\title{KeyedFileBrowser component}

\description{

}

\usage{
keyedFileBrowser(id=NULL, selectedPath=NULL, value=NULL)
}

\arguments{
\item{id}{Character. }

\item{selectedPath}{Character. }

\item{value}{Unnamed list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

