% Auto-generated: do not edit by hand
\name{keywords}

\alias{keywords}

\title{Keywords component}

\description{

}

\usage{
keywords(id=NULL, wrapperStyle=NULL, tagStyle=NULL, inputStyle=NULL,
tagDeleteStyle=NULL, placeholder=NULL, value=NULL)
}

\arguments{
\item{id}{Character. }

\item{wrapperStyle}{Named list. }

\item{tagStyle}{Named list. }

\item{inputStyle}{Named list. }

\item{tagDeleteStyle}{Named list. }

\item{placeholder}{Character. }

\item{value}{Unnamed list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

