% Auto-generated: do not edit by hand
\name{''DateTimePicker}

\alias{''DateTimePicker}

\title{DateTimePicker component}

\description{

}

\usage{
''DateTimePicker(id=NULL, datetime=NULL, timezone=NULL, value=NULL,
renderTimezone=NULL, style=NULL)
}

\arguments{
\item{id}{Character. }

\item{datetime}{Character. }

\item{timezone}{Character. }

\item{value}{Named list. }

\item{renderTimezone}{Logical. }

\item{style}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

