% Auto-generated: do not edit by hand
\name{dateTimePicker}

\alias{dateTimePicker}

\title{DateTimePicker component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
dateTimePicker(id=NULL, value=NULL, style=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{value}{Character. The value displayed in the input.}

\item{style}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

