% Auto-generated: do not edit by hand
\name{dateTimePicker}

\alias{dateTimePicker}

\title{DateTimePicker component}

\description{

}

\usage{
dateTimePicker(id=NULL, value=NULL, style=NULL)
}

\arguments{
\item{id}{Character. }

\item{value}{Character. }

\item{style}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

