% Auto-generated: do not edit by hand
\name{fileExplorer}

\alias{fileExplorer}

\title{FileExplorer component}

\description{

}

\usage{
fileExplorer(id=NULL, selectedPath=NULL, value=NULL)
}

\arguments{
\item{id}{Character. }

\item{selectedPath}{Character. }

\item{value}{Unnamed list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

