% Auto-generated: do not edit by hand
\name{keywords}

\alias{keywords}

\title{Keywords component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
keywords(id=NULL, wrapperStyle=NULL, tagStyle=NULL, inputStyle=NULL,
tagDeleteStyle=NULL, placeholder=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{wrapperStyle}{Named list. }

\item{tagStyle}{Named list. }

\item{inputStyle}{Named list. }

\item{tagDeleteStyle}{Named list. }

\item{placeholder}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

